requirements.txt
google-cloud-storage==2.10.0
bcrypt==4.0.1
absl-py==2.1.0
aniso8601==10.0.0
APScheduler==3.11.0
astunparse==1.6.3
bidict==0.23.1
blinker==1.9.0
boto3==1.37.26
botocore==1.37.26
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
DateTime==5.5
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==4.0.2
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.0
h11==0.16.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.3.1
pillow==11.0.0
protobuf==5.28.3
pyarrow==11.0.0
Pygments==2.18.0
pymongo==4.11.1
pymongoarrow==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s3transfer==0.11.4
simple-websocket==1.1.0
six==1.16.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3
urllib3==2.2.3
Werkzeug==3.0.6
wrapt==1.17.0
wsproto==1.2.0
xlsxwriter==3.2.5
zope.interface==7.2
gunicorn==21.2.0
flask-pymongo==2.3.0
PyJWT==2.8.0
google-cloud-storage==2.10.0
bcrypt==4.0.1